--- back/data/json/geo_test_1.JSON
@@ -1,10 +1,0 @@
-{
-  "type": "Feature",
-  "geometry": {
-    "type": "Point",
-    "coordinates": [125.6, 10.1]
-  },
-  "properties": {
-    "name": "Dingbat Islands"
-  }
-}
