--- back/data/json/test_1.JSON
@@ -1,5 +1,0 @@
-{
-  "name": "Jay",
-  "age": 21,
-  "education" : ["MSSM", "UMPI", "Brown"]
-}
