+++ back/src/main/java/edu/brown/cs/student/main/Server/Classes/Server.java
@@ -7,11 +7,8 @@ import edu.brown.cs.student.main.CSV.Classes.Parser;
 import edu.brown.cs.student.main.CSV.Classes.Searcher;
 import edu.brown.cs.student.main.CSV.Classes.Viewer;
 import edu.brown.cs.student.main.CSV.Interfaces.CSVDataSource;
-import edu.brown.cs.student.main.Server.Cache.RedlineCacheStrategy;
 import edu.brown.cs.student.main.Server.Interfaces.CensusDataSource;
 import spark.Spark;
-
-import java.util.ArrayList;
 
 public class Server {
 
@@ -26,7 +23,7 @@ public class Server {
 
   private final CSVDataSource csv;
 
-  public Server(CensusDataSource census, CSVDataSource csv, ArrayList<String> search_history) {
+  public Server(CensusDataSource census, CSVDataSource csv) {
     // Use whatever was dependency-injected into this constructor
     this.census = census;
     this.csv = csv;
@@ -42,10 +39,7 @@ public class Server {
     Spark.get("loadcsv", new Parser(this.csv));
     Spark.get("viewcsv", new Viewer(this.csv));
     Spark.get("searchcsv", new Searcher(this.csv));
-    Spark.get("broadband", new BroadbandHandler(this.census));
-    Spark.get("redline", new RedlineHandler(new RedlineCacheStrategy(5, 10)));
-    Spark.get("search-redline", new SearchRedlineHandler(new ArrayList<>()));
-    Spark.get("geobroadband", new GeoBroadbandHandler());
+    Spark.get("broadband", new CensusHandler(this.census));
 
     // Wait until the server has started.
     Spark.init();
@@ -55,7 +49,10 @@ public class Server {
   }
 
   public static void main(String[] args) {
-    Server server = new Server(new CensusAPI(), new CSVData(), new ArrayList<>());
+//    Use line 55 to run without cache. Comment line 55 and uncomment line 57 to run with cache.
+    Server server = new Server(new CensusAPI(), new CSVData());
+//    How we would have used server with caching functionality:
+//    Server server = new Server(new CensusCache(new CensusAPI()), new CSVData());
     System.out.println("Server started; exiting main...");
   }
 }
