--- back/src/main/java/edu/brown/cs/student/main/Server/Records/GeoBroadband.java
@@ -1,3 +1,0 @@
-package edu.brown.cs.student.main.Server.Records;
-
-public record GeoBroadband(Input input, Info[] results){}
