--- back/src/main/java/edu/brown/cs/student/main/Server/Records/Info.java
@@ -1,3 +1,0 @@
-package edu.brown.cs.student.main.Server.Records;
-
-public record Info(String state_name, String county_name){ }
