--- back/src/main/java/edu/brown/cs/student/main/Server/Records/Input.java
@@ -1,3 +1,0 @@
-package edu.brown.cs.student.main.Server.Records;
-
-public record Input(double lat, double lon, int censusYear){}
