--- back/src/main/java/edu/brown/cs/student/main/Server/Records/JsonData.java
@@ -1,7 +1,0 @@
-package edu.brown.cs.student.main.Server.Records;
-
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Objects;
-
-public record JsonData(Map<String, Object> data) { }
