+++ back/src/test/java/edu/brown/cs/student/Server/Tests/TestParser.java
@@ -7,9 +7,14 @@ import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.Moshi;
 import com.squareup.moshi.Types;
 import edu.brown.cs.student.Server.Mocks.MockedCSV;
+import edu.brown.cs.student.Server.Mocks.MockedCensusAPI;
 import edu.brown.cs.student.main.CSV.Classes.CSVData;
 import edu.brown.cs.student.main.CSV.Classes.Parser;
 import edu.brown.cs.student.main.CSV.Classes.Searcher;
+import edu.brown.cs.student.main.CSV.Interfaces.CSVDataSource;
+import edu.brown.cs.student.main.Server.Classes.CensusHandler;
+import edu.brown.cs.student.main.Server.Interfaces.CensusDataSource;
+import edu.brown.cs.student.main.Server.Records.CensusData;
 import edu.brown.cs.student.main.Server.Records.Geolocation;
 import java.io.IOException;
 import java.lang.reflect.Type;
@@ -20,6 +25,7 @@ import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import okio.Buffer;
+import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
